move api helper functions to _helpers
